+++ Craftsman/Models/ApiTemplate.cs
@@ -37,11 +37,6 @@ namespace Craftsman.Models
         public List<ApiEnvironment> Environments { get; set; } = new List<ApiEnvironment>();
 
         /// <summary>
-        /// Boolean that determines whether or not craftsman will do an initial git set up for this project.
-        /// </summary>
-        public bool AddGit { get; set; } = true;
-
-        /// <summary>
         /// The port that will be used when running locally in the project.
         /// </summary>
         public int Port { get; set; } = 5000;